Refs #22477 Moved class to workbench, and renamed widget to view
