+++ qt/applications/workbench/workbench/widgets/plotselector/model.py
