+++ qt/python/CMakeLists.txt
@@ -84,10 +84,6 @@ endif ()
     mantidqt/widgets/test/test_algorithmselector.py
     mantidqt/widgets/test/test_jupyterconsole.py
     mantidqt/widgets/test/test_messagedisplay.py
-
-    mantidqt/widgets/plotselector/test/test_plotselector_model.py
-    mantidqt/widgets/plotselector/test/test_plotselector_presenter.py
-    mantidqt/widgets/plotselector/test/test_plotselector_widget.py
   )
 
   # Tests
